+++ packages/layer-composer/src/generators/heatmap/modes/gridded.ts
@@ -50,6 +50,7 @@ export default function gridded(
             generatorType: Type.HeatmapAnimated,
             generatorId: config.id,
             interactive: true,
+            uniqueFeatureInteraction: true,
           },
         })
         chunkLayers.push({